Renamed BattleSystem folder to Battle, adjusted namespaces
