+++ Assets/Scripts/Runtime/Battle/UI/AttackPanelController.cs
@@ -1,10 +1,10 @@
 using System.Collections.Generic;
-using TMPro;
 using Jrpg.Core;
+using TMPro;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace Jrpg.Runtime.BattleSystem.UI
+namespace Jrpg.Runtime.Battle.UI
 {
     internal sealed class AttackPanelController : MonoBehaviour
     {
